==> code/0_Scraping_Project_3.ipynb <==
# Project 3 - Using Beautiful Soup, Reddit API and PushShift API to scrape SUBREDDITS

In this portion of the Project, we'll use [Beautiful Soup](https://www.crummy.com/software/BeautifulSoup/bs4/doc/), [Reddit API](https://www.reddit.com/dev/api/) and [Push Shift API](https://github.com/pushshift/api) to scrape the [ProCreate](https://www.reddit.com/r/ProCreate/) and [Adobe Illustrator](https://www.reddit.com/r/AdobeIllustrator/) subreddits.

We would discuss the pros and cons of the 3 different scraping mechanisms and our reasoning for picking PushShift API.

# Imports
import pandas as pd
import re
import requests
import time
import json
from time import sleep

# Import libaries
# imports
import pandas as pd
import requests
from bs4 import BeautifulSoup

## Contents

- [1.0 Using Beautiful Soup to Scrape](#using)
- [2.0 Utilize Reddits API to Scrape](#reddit)
- [3.0 Using PushShift API to Scrape](#push)

## 1.0 Using Beautiful Soup to Scrape<a name="using"></a>

Beautiful Soup is a Python library for pulling data out of HTML and XML files. It works with your favorite parser to provide idiomatic ways of navigating, searching, and modifying the parse tree. It commonly saves programmers hours or days of work.

Source: https://www.crummy.com/software/BeautifulSoup/bs4/doc/

# Create Beautiful SOUP function to scrape url page of REDDIT
def scraper_dictionary (url, subreddit):
    
    sodos = [] 
    
    dictlist = ['post', 'votes', 'comments','sub_reddit'] # essentially my header
    
    headers = {'User-Agent': 'Mozilla/5.0'}
    
    res = requests.get(url, headers=headers)
    
    soup = BeautifulSoup(res.content, 'html.parser')
    
    for i in soup.find_all('div',{'data-promoted':'false'}): # Ensures adverts are ignored
        
        empty = []
        
        # Posts
        y=i.find('a',{'data-event-action':'title'})
        empty.append(y.text)
    
        # Votes
        try:
            y=i.find('div',{'class':'score unvoted'})
            empty.append(int(y.text))
        except:
            continue
            
        # Comments
        y = i.find('a',{'data-event-action':'comments'})
        empty.append(y.text)
        
        empty.append(subreddit) # Subreddit
        
        sodos.append(empty)
        
    emptydict = []

    for i in sodos:

        emptydict.append(dict(zip(dictlist,i))) # This converts list of lists to dictionary 
# https://betterprogramming.pub/10-ways-to-convert-lists-to-dictionaries-in-python-d2c728d2aeb8

    return emptydict

# Create For Loop to loop through every 25 posts up to 1000 posts

final_dataframe = []

for i in range(0,1000,25):
     
    final_dataframe.extend(scraper_dictionary (f'https://old.reddit.com/r/ProCreate/?count={i}&after=t3_lrzhf4','Procreate'))
    
    sleep(11)

procreate = pd.DataFrame(final_dataframe)

> # 25
is the Maximum Number of Posts Beautiful Soup could scrape automatically due to Reddit's dynamic links which ensures each page link cannot be determined with a simple logic. 

procreate
# This did not work as Reddit utilizes random numbers to throw off webscraping page by page.
# So it just returned the same result 40 times.

# 2.0 Utilize Reddits API to Scrape<a name="reddit"></a>

An API to allow developers to build great products powered by Reddit because the developer community is integral to the success of the Reddit platform. The API is also to protect Reddit users’ privacy and security regardless of how they choose to consume Reddit content.

Source: https://www.reddit.com/wiki/api

import praw # This is a REDDIT Api created to Scrape Reddit pages officially

reddit = praw.Reddit(client_id='dD37vmEd4uZG0Q',client_secret='PEzHd6SVMQtz5FR1IvSuJGKANdgXSw',user_agent='project_3')
# I enter my reddit credentials

### Procreate New Posts
> This returns the latest 990 to 1000 posts in ProCreate subreddit

pc_new_posts = []

procreate = reddit.subreddit('ProCreate')

for pc_post in procreate.new(limit = 1000):
    
    pc_new_posts.append([pc_post.title, pc_post.score, pc_post.id, pc_post.subreddit, pc_post.url, pc_post.num_comments, pc_post.selftext, pc_post.created])
    
pc_new_posts = pd.DataFrame(pc_new_posts,columns=['title', 'score', 'id', 'subreddit', 'url', 'num_comments', 'body', 'created'])

len(pc_new_posts)

### Illustrator New Posts

> This returns the latest 990 to 1000 posts in Adobe Illustrator subreddit

ai_new_posts = []

ai = reddit.subreddit('AdobeIllustrator')

for post in ai.new(limit = 1000):
    
    ai_new_posts.append([post.title, post.score, post.id, post.subreddit, post.url, post.num_comments, post.selftext, post.created])
    
ai_new_posts = pd.DataFrame(ai_new_posts,columns=['title', 'score', 'id', 'subreddit', 'url', 'num_comments', 'body', 'created'])

len(ai_new_posts)

### Procreate Top Posts

> This returns the top 990 to 1000 posts in ProCreate subreddit

pc_top_posts = []

procreate = reddit.subreddit('ProCreate')

for pc_post in procreate.top(limit = 1000):
    
    pc_top_posts.append([pc_post.title, pc_post.score, pc_post.id, pc_post.subreddit, pc_post.url, pc_post.num_comments, pc_post.selftext, pc_post.created])
    
pc_top_posts = pd.DataFrame(pc_top_posts,columns=['title', 'score', 'id', 'subreddit', 'url', 'num_comments', 'body', 'created'])

len(pc_top_posts)

### Adobe Illustrator Top Posts

> This returns the top 990 to 1000 posts in Adobe Illustrator subreddit

ai_top_posts = []

ai = reddit.subreddit('AdobeIllustrator')

for post in ai.top(limit = 1000):
    
    ai_top_posts.append([post.title, post.score, post.id, post.subreddit, post.url, post.num_comments, post.selftext, post.created])
    
ai_top_posts = pd.DataFrame(ai_top_posts,columns=['title', 'score', 'id', 'subreddit', 'url', 'num_comments', 'body', 'created'])

len(ai_top_posts)

### Procreate Hot Posts

> This returns the hottest 990 to 1000 posts in ProCreate subreddit

pc_hot_posts = []

procreate = reddit.subreddit('ProCreate')

for pc_post in procreate.top(limit = 1000):
    
    pc_hot_posts.append([pc_post.title, pc_post.score, pc_post.id, pc_post.subreddit, pc_post.url, pc_post.num_comments, pc_post.selftext, pc_post.created])
    
pc_hot_posts = pd.DataFrame(pc_hot_posts,columns=['title', 'score', 'id', 'subreddit', 'url', 'num_comments', 'body', 'created'])

len(pc_hot_posts)

### Ilustrator Hot Posts

> This returns the top 990 to 1000 posts in Aodbe Illustrator subreddit

ai_hot_posts = []

ai = reddit.subreddit('AdobeIllustrator')

for post in ai.top(limit = 2000):
    
    ai_hot_posts.append([post.title, post.score, post.id, post.subreddit, post.url, post.num_comments, post.selftext, post.created])
    
ai_hot_posts = pd.DataFrame(ai_hot_posts,columns=['title', 'score', 'id', 'subreddit', 'url', 'num_comments', 'body', 'created'])

len(ai_hot_posts)

ai_hot_posts.head() # Inspect Posts

> # 1000
is the Maximum Number of Posts the Reddit API could scrape automatically due to Reddit's limit on its API. It was by far the simplest of the 3 scraping mechanisms to use but it did not give me enough data. My target was at least 5,000

ai_top_posts.head() # Inspect Posts

ai_new_posts.head() # Inspect Posts

## 3.0 Using PushShift API to Scrape<a name="push"></a>

The pushshift.io Reddit API was designed and created by the /r/datasets mod team to help provide enhanced functionality and search capabilities for searching Reddit comments and submissions. The project lead, /u/stuck_in_the_matrix, is the maintainer of the Reddit comment and submissions archives located at https://files.pushshift.io.

This RESTful API gives full functionality for searching Reddit data and also includes the capability of creating powerful data aggregations. With this API, you can quickly find the data that you are interested in and find fascinating correlations.

Source: https://github.com/pushshift/api

def subreddit_scrape (subreddit,start_date):
    
    url = 'https://api.pushshift.io/reddit/submission/search/' # My base link
    
    empty = [] # Empty string to tag on the Scraped Data
    
    while (start_date < 1614314325): # Set an Epoch Time Limit

        params = {
            'subreddit': subreddit,
            'size': 100,
            'after': start_date,
            'sort_type' : 'created_utc',
            'fields' : ['subreddit','title','selftext','created_utc','score','num_comments','subreddit_subscribers','full_link']
        } # Define Parameters to scrape from Reddit

        res = requests.get (url, params) # Combine parameters with URL

        dataseries = res.json() # Extract .json data from URL page
        
        empty.extend(dataseries['data']) # Add Data to empty list
        
        sleep(3) # Sleep for 3 seconds so we don't get flagged by Reddit
        
        if len(dataseries['data']) == 100: # This is to ensure there is no error 
        #occurs if less than 100 posts returned
        
            start_date = dataseries['data'][99]['created_utc']
            # Tag on the oldest date to our list and continue iterating
            
        else:
            
            break
        
    df = pd.DataFrame(empty) # Convert lists to dataframe

    return df

# https://pushshift.io/api-parameters/

procreate = subreddit_scrape ('ProCreate',1546306316) # Start Jan 2019

> Export PushShift API dataset to our Data Folder

# Write the DataFrame you created to a csv called 'procreate_finals.csv'
procreate.to_csv('data/0_Scraped_Data/procreate_final.csv', index=False)
print('Submission CSV is ready!')

> # 28,000!
Is the maximum number of posts we were able to scrape with the PushShift API and this was for the ProCreate Reddit. It took the longest to execute of the three, but the reward for effort was worth it.

procreate.columns # Inspect Columns

procreate.shape # Inspect Shape

procreate[procreate['num_comments'] <= 2].shape # Check posts with comments less than 2

procreate[procreate['selftext'] == '[deleted]'].shape # Check posts with deleted self text

adobeillustrator = subreddit_scrape('AdobeIllustrator',1546306316) # Start Jan 2019

> Export PushShift API dataset to our Data Folder

# Write the DataFrame you created to a csv called 'adobeillustrators.csv'
adobeillustrator.to_csv('data/0_Scraped_Data/adobeillustrator.csv', index=False)
print('Submission CSV is ready!')

adobeillustrator.shape # Inspect Shape

adobeillustrator[adobeillustrator['num_comments'] <= 2].shape

adobeillustrator[adobeillustrator['selftext'] == '[deleted]'].shape

“I am feeling 😊 today"
