==> code/1_Cleaning_Data_Project_3.ipynb <==
# Project 3 - Concating SubREDDITS and doing Preliminary Cleaning

In this portion of the Project, we'll concat both subreddits scraped data and trim it down to focus on the more active subreddit posts, which are the ones with a good amount of comments. We will also create 2 new columns for further analysis down the line.

# Data Manipulation
import numpy as np
import pandas as pd

# this setting widens how many characters pandas will display in a column:
pd.options.display.max_colwidth = 350

pr_df =  pd.read_csv('data/0_Scraped_Data/procreate_final.csv')

ai_df = pd.read_csv('data/0_Scraped_Data/adobeillustrator.csv')

pr_df.head(1)

ai_df.head(1)

# Trim rows down

pr_df = pr_df[(pr_df['num_comments'] > 4) & (pr_df['selftext'] != '[deleted]')]

ai_df = ai_df[(ai_df['num_comments'] > 4) & (ai_df['selftext'] != '[deleted]')]

project_ai_df = pd.DataFrame(data=ai_df, columns = ['subreddit','title'])

project_pr_df = pd.DataFrame(data=pr_df, columns = ['subreddit','title'])

project_ai_df.head(1)

project_pr_df.head(1)

frames = [project_pr_df, project_ai_df]
ai_pr_df = pd.concat(frames, axis = 0) # Concatenate the dataframes row to row

ai_pr_df.head(1)

ai_pr_df.tail(1)

ai_pr_df.reset_index(drop=True, inplace = True)

### Create DataFrame Column With Our Time

frames = [pr_df, ai_df]
ai_pr_time = pd.concat(frames, axis = 0) # Concatenate the dataframes row to row

ai_pr_time.head(1)

ai_pr_time.drop(['full_link','num_comments','score','selftext','subreddit'], axis=1, inplace=True)

ai_pr_time.head(1)

# Write the DataFrame you created to a csv called 'predictions.csv'
ai_pr_time.to_csv('data/1_Scraped_Data_IBM/ibm_watson_time.csv', index=False)
print('Submission CSV is ready!')

### Create a new column called `status_char_length` that contains the character length of each status

> Note: You can do this in one line with `map`.

## Title lenghts
ai_pr_df['status_char_length'] = [len(ai_pr_df['title'][i]) for i in range(0,ai_pr_df['title'].shape[0])]

### Create a new column called `status_word_count` that contains the number of words in each status

> Note: You can evaluate this based off of how many strings are separated by whitespaces; you're not required to check that each set of characters set apart by whitespaces is a word in the dictionary.

ai_pr_df['status_word_count'] = [len(ai_pr_df['title'][i].split()) for i in range(0,ai_pr_df['title'].shape[0])] # Split on the spaces and then count the number of words with Lenght in-built function

ai_pr_df.head()

# Write the DataFrame you created to a csv called 'predictions.csv'
ai_pr_df.to_csv('data/1_Scraped_Data_IBM/ibm_watson.csv', index=False)
print('Submission CSV is ready!')

“I am feeling 🤔 today"
