==> code/2_IBM_Sentiment_Analysis.ipynb <==
# Project 3 - Using Natural Language Understanding and Sentiment

> <font color=red>**NOTE:** This Jupyter Notebook is a save-as from an Existing IBM NoteBook from the GitHub of Scott D'Angelo with minor modifications made by me. Also the order of flow of the NoteBook will be off, due to API call limitations by IBM Watson NLU Program. This meant I could not rerun the notebook to get the code to be chronological.</font>

In this portion of the workshop, we'll use an instance of [Watson Natural Language Understanding](https://cloud.ibm.com/catalog/services/natural-language-understanding) to gather insights into data.

Watson Natural Language Understanding is a cloud native product that uses deep learning to extract metadata from text such as entities, keywords, categories, sentiment, emotion, relations, and syntax.
There is a rich [API](https://cloud.ibm.com/apidocs/natural-language-understanding?code=python) that we will use along with the [Watson Python SDK](https://github.com/watson-developer-cloud/python-sdk) to analyze our data.

## Contents

- [1.0 Setup - install modules](#setup)
- [2.0 Test NLU APIs](#test)
- [3.0 Import Data and Setup Pandas Dataframe ](#pandas)
- [4.0 Clean and Prepare Data for NLU scoring](#clean)
- [5.0 Get Sentiment by Row](#sentiment-row)

## 1.0 Setup - Install Modules<a name="setup"></a>

We use the [Watson Python SDK](https://github.com/watson-developer-cloud/python-sdk) to access the [NLU APIs](https://cloud.ibm.com/apidocs/natural-language-understanding?code=python) programatically.

Import python modules from the Watson Python SDKs

import json
from ibm_watson import NaturalLanguageUnderstandingV1
from ibm_cloud_sdk_core.authenticators import IAMAuthenticator
from ibm_watson.natural_language_understanding_v1 import Features,CategoriesOptions,EmotionOptions,KeywordsOptions

# Warning
import warnings
warnings.filterwarnings('ignore')

import pandas as pd
import numpy as np
from datetime import datetime

### 1.1 Add NLU credentials
Get the [IAM Authentication Key](https://cloud.ibm.com/apidocs/natural-language-understanding?code=python#authentication) and [Service URL](https://cloud.ibm.com/apidocs/natural-language-understanding?code=python#service-endpoint) that you obtained when you [Created a Watson NLU instance](https://github.ibm.com/IBMDeveloper/python-and-analytics/tree/addNLU/workshop/natural-language-understanding#create-a-watson-nlu-instance).

Add your [IAM Authentication Key](https://cloud.ibm.com/apidocs/natural-language-understanding?code=python#authentication) below.

IAM_KEY = ''

Add your [NLU Service URL](https://cloud.ibm.com/apidocs/natural-language-understanding?code=python#service-endpoint) below

SERVICE_URL = 'https://api.us-south.natural-language-understanding.watson.cloud.ibm.com/instances/'

## 2.0 Test NLU APIs <a name="test"></a>
Run a quick check to make sure everything is working. We'll use a [basic web page](https://www.ibm.com) to see how Watson Natural Language Understanding can extract categories when given a URL. [This example](https://cloud.ibm.com/apidocs/natural-language-understanding?code=python#categories) comes from the Watson NLU documentation.

authenticator = IAMAuthenticator(IAM_KEY)
natural_language_understanding = NaturalLanguageUnderstandingV1(version='2020-08-01',authenticator=authenticator)

natural_language_understanding.set_service_url(SERVICE_URL)

response = natural_language_understanding.analyze(
    url='www.ibm.com',
    features=Features(categories=CategoriesOptions(limit=3))).get_result()

print(json.dumps(response, indent=2))

## 3.0 Import Data and Setup Pandas Dataframe<a name="pandas"></a>

df = pd.read_csv('/data/1_Scraped_Data_IBM/ibm_watson.csv')
df.head(5)

## 4.0 Clean and Prepare data for NLU scoring<a name="clean"></a>

Now, let's look for something that we can use with Watson NLU to derive an analysis of the sentiment of the customer feedback.

df[df['status_word_count'] > 5]['subreddit'].value_counts()

df_five = df[df['status_word_count'] > 5]

df_five.reset_index(inplace = True, drop = True)

df_five.head(1357)

print (df_five['title'][0])

That looks like what we want. Now, we'll create a list to hold the `responses`, call Watson NLU with the data and then populate the responses list. We'll do the same with a list called `normalize` that we can use along with [json_normalize()](https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.json_normalize.html).

## 5.0 Get Sentiment by Row<a name="sentiment-row"></a>
Now, let's derive some sentiment and emotion information on a per-row basis, to provide more granualarity.
The number of API calls that you can make to Watson NLU is [rate limited and dependent on your service plan](https://cloud.ibm.com/catalog/services/natural-language-understanding).

responses = []
normalize = []
for index, row in df_five.iterrows():
    
    response = natural_language_understanding.analyze(
    text = row['title'],
        language = 'en',
    features=Features(keywords=KeywordsOptions(sentiment=True,emotion=True,limit=1))).get_result()
    normalize.append(pd.json_normalize(response['keywords']))
    responses.append(response)

Add the `responses` list and the `normalize` to the df_rows dataframe. We can continue to use these new data features, but more commonly we'll derive new dataframes for our experiments and change those new dataframes instead.

warnings.filterwarnings('ignore')
df_five['response'] = responses
df_five.head()

warnings.filterwarnings('ignore')
df_five['normalized'] = normalize
df_five.head()

# Write the DataFrame you created to a csv called 'predictions.csv'
df_five.to_csv('/data/2_Cleaned_IBM_Data/cleaned_ibm.csv', index=False)
print('Submission CSV is ready!')

“I am feeling 😕   today"
